Pull HTTP subroutines out of the main module
Also add ISIError exception (tsk, shoulda been in a separate commit)
